toggle details works now yaaay
toggle function finally working. Now one can click on one element and open and close it in ListForm but still keep the grid-form intact. Solved it by adding an if not listForm before the if else statement that render out the specific listStyle to be able to expand and close the chosen div