continue field form add researcher
trying to make radiobutton or checkbox in form so that more working groups and region of expertise can be added to avoid unnecessary fields. Also trying to make a function to render out the options of year as well as the other options to avoid hardcoding everything. Not yet succeeded. After that, it should all go to the jsonserver